Add files via upload
The program reads user-defined Meta_File.csv and data files downloaded by get_EPA_files.sh script. Then it sorts the data and makes an output file
in .mat format. It uses 2 functions: one for  Wind data extraction and sorting and the other for the custom variable(s). A detailed description is in the program
and functions file.